move description and plot description
