+++ src/pandas_profiling/model/description_plot.py
@@ -1,0 +1,116 @@
+from dataclasses import dataclass
+from typing import Optional
+
+import numpy as np
+import pandas as pd
+
+
+@dataclass
+class BasePlotDescription:
+    """Base class for plot description.
+
+    Attributes
+    ----------
+    data_col_name : str
+        Name of data column.
+    target_col_name : str or None
+        Name of target column.
+    count_col_name : str
+        Name of count column in preprocessed DataFrames.
+    log_odds_col_name : str
+        Name of log2odds column in log_odds DataFrame.
+    distribution : pd.DataFrame
+        Distribution DataFrame preprocessed for plotting.
+    log_odds : pd.DataFrame or None
+        Log2odds DataFrame preprocessed for plotting.
+    """
+
+    data_col_name: str
+    target_col_name: Optional[str]
+
+    __distribution: pd.DataFrame
+    __log_odds: Optional[pd.DataFrame]
+
+    count_col_name: str = "count"
+    log_odds_col_name: str = "log_odds"
+
+    def __init__(self, data_col_name: str, target_col_name: Optional[str]) -> None:
+        """Setup column names for data_col and target_col.
+
+        Parameters
+        ----------
+        data_col_name: str
+            Name of data column.
+        target_col_name: str or None
+            Name of target column.
+        """
+        self.data_col_name = data_col_name
+        self.target_col_name = target_col_name
+
+    @property
+    def distribution(self) -> pd.DataFrame:
+        """Returns preprocessed DataFrame for plotting
+
+        distribution: pd.DataFrame with 2 or 3 columns (data_col, target_col or None, count)
+        in format:
+            col_name,   target_name,    count
+            1           0               10
+            1           1               5
+            2           0               8
+            ..."""
+        if self.__distribution is None:
+            raise ValueError(
+                "Distribution not set at '{}' variable.".format(self.data_col_name)
+            )
+        return self.__distribution
+
+    @property
+    def log_odds(self) -> Optional[pd.DataFrame]:
+        """Returns DataFrame with relative log2odds for data column.
+        format:
+            col_name,   log_odds
+            male        -2
+            female      2
+        """
+        return self.__log_odds
+
+    def __generate_log_odds(self):
+        """Generates log2odds preprocessed DataFrame based on distribution."""
+        log_odds = pd.pivot_table(
+            self.distribution,
+            values=self.count_col_name,
+            index=self.data_col_name,
+            columns=self.target_col_name,
+            sort=False,
+        ).reset_index()
+        log_odds.columns.name = ""
+        # TODO replace '0' and '1'
+        log_odds["log_odds"] = round(np.log2(log_odds["1"] / log_odds["0"]), 2)
+        # replace all special values with 0
+        log_odds.fillna(0, inplace=True)
+        log_odds.replace([np.inf, -np.inf], 0, inplace=True)
+        self.__log_odds = log_odds
+
+    def _set_distribution(self, distribution: pd.DataFrame) -> None:
+        """Validate and set distribution DataFrame."""
+        if not isinstance(distribution, pd.DataFrame):
+            raise ValueError("Preprocessed plot must be pd.DataFrame instance.")
+        self.__check_columns(distribution)
+        self.__distribution = distribution.reset_index(drop=True)
+
+        # generate log_odds just for supervised report
+        if self.target_col_name is not None:
+            self.__generate_log_odds()
+
+    def __check_columns(self, df: pd.DataFrame):
+        """Checks if df contains all columns (data_col, target_col, count_col)."""
+        if self.data_col_name not in df:
+            raise ValueError(
+                "Data column '{}' not in DataFrame.".format(self.data_col_name)
+            )
+        if (self.target_col_name is not None) and (self.target_col_name not in df):
+            raise ValueError(
+                "Target column '{}' not in DataFrame.".format(self.target_col_name)
+            )
+        if self.count_col_name not in df:
+            raise ValueError("Count column not in DataFrame.")
